+++ packages/core/src/server/middlewares/lazyCompilation.ts
@@ -8,6 +8,7 @@ import {
   bold,
   cyan,
   getDynamicResources,
+  getValidPublicPath,
   green
 } from '../../index.js';
 import { send } from '../send.js';
@@ -21,7 +22,10 @@ export function lazyCompilationMiddleware(
 ): Connect.NextHandleFunction {
   return async function handleLazyCompilationMiddleware(req, res, next) {
     const { config, compiler } = app;
-    const lazyCompilationPath = `/${DEFAULT_LAZY_COMPILATION_PATH}`;
+    const publicPath = getValidPublicPath(
+      compiler.config?.compilation?.output?.publicPath
+    );
+    const lazyCompilationPath = `${publicPath || '/'}${DEFAULT_LAZY_COMPILATION_PATH}`;
 
     if (!req.url.startsWith(lazyCompilationPath)) {
       return next();