+++ examples/tailwind-next/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @farmfe-examples/tailwind-next
 
+## 0.0.21
+
+### Patch Changes
+
+- @farmfe/js-plugin-tailwindcss@0.0.20
+
+## 0.0.20
+
+### Patch Changes
+
+- @farmfe/js-plugin-tailwindcss@0.0.19
+
 ## 0.0.19
 
 ### Patch Changes